requirements.txt
fpdf==1.7.2
Pillow==10.1.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pytest==7.4.3




